+++ libs/unity-py/CHANGELOG.md
@@ -5,8 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2025-02-19
+
+### Added
+
+### Fixed
+
+### Changed
+* Changed Health Service to report information from updated Health API endpoint.
+* Minor changes to the Health Service Status report.
+* Updated health JSON schema definition to match updated output from Health API endpoint.
+
+### Removed
+
+### Security
+
+### Deprecated
 
 ## [0.8.1] - 2025-02-10
+
 ### Added
 
 ### Fixed